refactor: /community/[boardCode]/[postId]/modify 페이지 전환
